backend/requirements.txt
# Django Framework
django>=5.1.1

django-cors-headers==4.0.0

djangorestframework==3.14.0

# PostgreSQL Database
psycopg2-binary>=2.9.1

# Google APIs for Gmail and Calendar
google-api-python-client>=2.60.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6

# OpenAI API for AI integration
openai

# Timezone handling
tzlocal>=4.1

# OAuth 2.0 authentication
google-auth>=2.0.0

# JSON Web Token (optional, if used for authentication)
PyJWT>=2.1.0

# Date and Time utilities
python-dateutil>=2.8.2

# Parsing and manipulating JSON
jsonschema>=4.2.1

# Email utilities
httplib2>=0.20.0

# Utility for reading environment variables
python-dotenv>=0.19.0

# Pandas for data manipulation
pandas>=1.3.0

# Scikit-learn for model selection and preprocessing
scikit-learn>=1.0.0

# Transformers library for BERT tokenizer
transformers>=4.0.0

# PyTorch for deep learning
#torch>=1.9.0

# NumPy for numerical operations
numpy>=1.21.0

unidecode
#ntlk
nrclex
deep-translator
# redis-server
ldaphne
langchain
langchain-openai
daphne
channels
jinja2
channels_redis